Fixes for layout
There's quite a few changes here:
Firstly I deleted the top level 'index.html' and 'style.css' and 'images', and replaced them with the versions that were in the subdirectory 'J-b172.github.io-main, and then deleted that subdirectory.  This is because github pages will only look for the index.html in the top directory.

 I also rescaled the images to suit the desired size in the desktop layout. There's no good reason to have images stored at a bigger size than that, and it just makes the CSS sizing tricky if there are very large images it is trying to scale down.

There was some confusion between 'id' and 'class' in the html particularly with the story cards (and their images and banners).  A 'class' is something that you can apply to several elements at once (for example story-card-banner) whilst an 'id' is unique to a particular element (for example storyCardBanner1).  In the css you can apply a style to an entire class of elements using the prefix "." whereas you can apply a style to a particular id using the prefix "#"

Also the previous code was attempting to use the grid layout to place the story card banners, whilst simultaneously having these banners as child elements of the story-card, and giving the story-card a flex layout. You can't do both of these things at once.  If you want to place an element in the grid, then it needs to be a direct child element of the grid container (which in this case is the mainContainer element)